fix(execution): unblock auto-trading — 3 blockers removed
1. votes_for_action int→float in SignalResponse (CRITICAL)
   Pydantic's strict int validator rejected fractional weighted votes
   (e.g. 17.4 for LLY HOT BUY) with int_from_float ValidationError →
   HTTP 500 → orchestrator logged "Brain API call failed" and skipped
   every BUY/SELL signal. Only signals with whole-number vote totals
   (e.g. DOGEUSD=14.0) slipped through. Changed to float=0.0 to match
   signal.py and the weighted-voting system.

2. Orchestrator log format %d→%.1f for votes_for_action
   Silently truncated float vote counts in logs (cosmetic but misleading).

3. Gate 5b: skip crypto SELL when no open position exists
   Alpaca does not support crypto shorting. Without this gate, every
   crypto SELL signal with no position triggered a wasted /execute call
   returning 409. Now caught early with a clear log message.

+++ brain/api.py
@@ -205,9 +205,9 @@ class SignalResponse(BaseModel):
     take_profit_pct: float
     agent_views: dict[str, str]
     passed_confidence_gate: bool
-    # Vote-based fields — combined 15-agent pool
+    # Vote-based fields — combined 27-agent weighted pool
     vote_tally: dict = {}
-    votes_for_action: int = 0
+    votes_for_action: float = 0.0
     regime_label: str = "UNKNOWN"
     tier: str = "WARM"
     devil_advocate_score: int = 0
